+++ src/monster.h
@@ -38,13 +38,18 @@ struct MonsterCost {
 Monster* monster_initialize(Dungeon* dungeon, u_char x, u_char y, u_char floor, bool canTunnel);
 Monster* monster_terminate(Monster* monster);
 
-int32_t monster_path_cmp(const void* key, const void* with);
+int monster_run_dijkstra_on_all_floors(Dungeon* dungeon);
+int monster_run_dijkstra_on_floor(Floor* floor);
+
+int32_t monster_dijkstra_compare(const void* A, const void* B);
 void monster_dijkstra_tunneler(Floor* floor);
 void monster_dijkstra_non_tunneler(Floor* floor);
 
-void monster_free_standing_on(Monster* monster);
+int monster_free_dungeon(Dungeon* dungeon);
+int monster_free_floor(Floor* floor);
+int monster_free(Monster* monster);
 
-void monster_move_to(Monster* monster, u_char toX, u_char toY);
-void monsters_move(Floor* floor);
+int monster_move_to(Monster* monster, u_char toX, u_char toY);
+int monsters_move(Floor* floor);
 
 #endif