+++ src/player.h
@@ -22,7 +22,7 @@ struct Player {
 Player* player_initialize(Dungeon* dungeon, u_char x, u_char y, u_char floor);
 Player* player_terminate(Player* player);
 
-void player_free_standing_on(Player* player);
-void player_move_to(Player* player, u_char toFloor, u_char toX, u_char toY);
+int player_free(Player* player);
+int player_move_to(Player* player, u_char toFloor, u_char toX, u_char toY);
 
 #endif