+++ src/save_load.h
@@ -11,10 +11,10 @@
 
 int save_to_file(Dungeon* dungeon);
 int load_from_file(Dungeon* dungeon);
+int load_from_program(Dungeon* dungeon);
 int save_floor(FILE* file, Dungeon* dungeon, u_char floorIndex);
 int load_floor(FILE* file, Dungeon* dungeon, u_char floorIndex);
 
 void save_error(Dungeon* dungeon);
-void load_error();
 
 #endif