+++ src/settings.h
@@ -1,6 +1,9 @@
 #ifndef SETTINGS_H
 #define SETTINGS_H
 
+struct Settings;
+typedef struct Settings Settings;
+
 #include <stdbool.h>
 #include <libgen.h>
 #include <dirent.h>
@@ -8,14 +11,13 @@
 #include <sys/types.h>
 #include "resource.h"
 
-struct Settings;
-typedef struct Settings Settings;
-
 #define RESOURCE_PATH "./res"
 #define DATA_PATH "/.rlg327/"
 #define FILE_HEADING "RLG327-S2019"
 
 struct Settings {
+    bool expandedPrint;
+
     bool doSave;
     bool doLoad;
 
@@ -28,7 +30,7 @@ struct Settings {
 Settings* settings_initialize(int argc, char* argv[]);
 Settings* settings_terminate(Settings* settings);
 
-void create_home_folders();
-int load_function_arguments(Settings* settings, int argc, char* argv[]);
+int create_home_folders();
+int settings_load_arguments(Settings* settings, int argc, char* argv[]);
 
 #endif